rename species names module (pkm_rs_resource/src/pkhex_text)
