+++ pkm_rs_resources/src/pkhex_text/mod.rs
@@ -1,3 +1,3 @@
-mod names;
+mod species_names;
 
-pub use names::{species_name, species_name_en};
+pub use species_names::{species_name, species_name_en};
