+++ pkm_rs_resources/src/pkhex_text/species_names.rs
